requirements.txt
streamlit
pandas
altair
groq
google-api-python-client
feedparser
requests
python-dotenv